Setup push notification for LMS Enrollment Request
